requirements.txt
PyQt5
ffmpeg-python

